+++ tests/codeception/_support/AcceptanceTester.php
@@ -1,0 +1,19 @@
+<?php
+
+/**
+ * Acceptance tester class.
+ *
+ * @package WordPoints_WooCommerce
+ * @since 1.2.0
+ */
+
+/**
+ * Tester for use in the acceptance tests.
+ *
+ * @since 1.2.0
+ */
+class AcceptanceTester extends \WordPoints\Tests\Codeception\AcceptanceTester {
+	// Add any methods here that you want.
+}
+
+// EOF